change int in size_t

+++ srcs/ft_strchr.c
@@ -6,7 +6,7 @@
 /*   By: lowatell <lowatell@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/30 00:54:16 by lowatell          #+#    #+#             */
-/*   Updated: 2024/10/08 10:51:40 by lowatell         ###   ########.fr       */
+/*   Updated: 2024/10/15 23:13:58 by lowatell         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strchr(const char *s, int c)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
